Move the stub data construction to services
